% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eignor.R
\name{eignor}
\alias{eignor}
\title{Maximum eigen value normalisation}
\usage{
eignor(W)
}
\arguments{
\item{W}{spatial weights matrix}
}
\value{
\describe{\emph{W}}  Eigen normalised spatial weights matrix
}
\description{
Maximum eigen value row normalisation of a spatial weight matrix.
}
\examples{
data(gN3dist)
dist2<-gN3dist/1000 ##distance in km
W<-InvDistMat(distMat = dist2, distCutOff = 100, powr = 2)
Wnor<-eignor(W)

}
\seealso{
\code{\link{rownor}}
}
\author{
Rozeta Simonovska
}
